Add patho-mvp version
There's a loooot of refactoring necessary, but I just
wanted to have the working version commited. Also,
removed that extension based version of Flux bootstrap.
It installed an old version of Flux and soon it will be
replaced with the deployment script approach. This will
allow to contain the whole AKS deployment and bootstrap
inside Bicep without relying on GitHub workflows.